done topup-storage
added:
1. files related to topup_storage
2. func for float addition with defined percision

modified:
1. the order of getting the latest inserted harga ID and querying the DB to guarantee that cek harga will always return the newest harga

+++ model/rekening.go
@@ -2,6 +2,6 @@ package model
 
 type Rekening struct {
 	Norek string  `json:"norek,omitempty"`
-	Saldo float32 `json:"saldo,omitempty"`
+	Saldo float64 `json:"saldo,omitempty"`
 	TmplCols
 }
